License fixed, hierarchy simplifications
- Agent generation moved to a single file, moved from main directory.
- Initializers fixed.
- Environment imports reordered.
