+++ RouteRL/environment/__init__.py
@@ -2,6 +2,9 @@ from .agent import BaseAgent
 from .agent import HumanAgent
 from .agent import MachineAgent
 
+from .agent_generation import create_agent_objects
+from .agent_generation import generate_agents_data
+
 from .simulator import SumoSimulator
 
 from .environment import TrafficEnvironment
