+++ RouteRL/network_and_config/agents_data.csv
@@ -1,101 +1,101 @@
 id,origin,destination,start_time,kind
 0,1,1,37,Human
-1,1,1,29,Human
-2,0,1,23,Human
+1,1,0,29,Human
+2,1,0,23,Human
 3,1,1,25,Human
 4,0,1,45,Human
-5,0,0,14,Human
-6,0,0,23,Human
-7,0,0,21,Human
+5,1,0,14,Human
+6,1,1,23,Human
+7,1,0,21,Human
 8,1,1,24,Human
-9,1,1,23,Human
-10,0,0,20,Human
-11,0,0,25,Human
+9,1,0,23,Human
+10,1,0,20,Human
+11,1,0,25,Human
 12,0,1,18,Human
-13,1,0,28,Human
-14,0,1,12,Human
-15,1,1,20,Human
-16,0,0,20,Human
-17,1,1,19,Human
-18,0,1,27,Human
+13,0,0,28,Human
+14,0,0,12,Human
+15,0,0,20,Human
+16,0,1,20,Human
+17,0,1,19,Human
+18,0,0,27,Human
 19,0,1,20,Human
-20,0,1,25,Human
-21,0,0,25,Human
-22,0,1,14,Human
-23,1,1,24,Human
-24,0,0,21,Human
-25,1,0,22,Human
-26,1,0,27,Human
-27,0,1,23,Human
-28,0,0,20,Human
+20,1,0,25,Human
+21,1,0,25,Human
+22,1,1,14,Human
+23,1,0,24,Human
+24,1,0,21,Human
+25,0,0,22,Human
+26,0,1,27,Human
+27,0,0,23,Human
+28,1,0,20,Human
 29,0,1,17,Human
-30,0,1,33,Human
-31,0,1,15,Human
-32,1,1,27,Human
-33,1,0,45,Human
-34,1,0,9,Human
+30,1,1,33,Human
+31,1,0,15,Human
+32,0,1,27,Human
+33,0,0,45,Human
+34,0,0,9,Human
 35,0,0,34,Human
 36,1,0,23,Human
-37,1,0,44,Human
+37,1,1,44,Human
 38,0,1,30,Human
-39,1,0,28,Human
+39,0,0,28,Human
 40,1,0,24,Human
-41,0,0,28,Human
+41,1,0,28,Human
 42,0,0,22,Human
-43,1,1,40,Human
-44,0,1,30,Human
+43,0,1,40,Human
+44,1,1,30,Human
 45,1,1,32,Human
-46,1,1,30,Human
-47,0,1,27,Human
-48,1,1,4,Human
-49,1,0,21,Human
-50,1,0,17,Human
-51,1,0,9,Human
-52,0,1,32,Human
-53,1,1,15,Human
-54,0,0,34,Human
-55,0,1,26,Human
-56,1,1,26,Human
-57,0,1,35,Human
-58,1,1,24,Human
-59,1,1,28,Human
-60,0,0,28,Human
-61,1,1,20,Human
-62,1,0,35,Human
+46,1,0,30,Human
+47,1,1,27,Human
+48,0,0,4,Human
+49,0,1,21,Human
+50,0,1,17,Human
+51,0,0,9,Human
+52,1,1,32,Human
+53,0,1,15,Human
+54,0,1,34,Human
+55,1,0,26,Human
+56,1,0,26,Human
+57,1,0,35,Human
+58,0,1,24,Human
+59,1,0,28,Human
+60,1,1,28,Human
+61,0,0,20,Human
+62,0,0,35,Human
 63,0,0,14,Human
-64,0,0,8,Human
-65,0,1,21,Human
-66,1,0,21,Human
+64,0,1,8,Human
+65,1,0,21,Human
+66,0,1,21,Human
 67,0,1,17,Human
-68,1,1,36,Human
-69,1,0,22,Human
+68,0,0,36,Human
+69,0,0,22,Human
 70,0,1,36,Human
-71,0,0,28,Human
-72,0,1,23,Human
-73,0,0,22,Human
+71,0,1,28,Human
+72,0,0,23,Human
+73,1,0,22,Human
 74,1,1,23,Human
-75,0,0,27,Human
-76,1,0,11,Human
+75,1,0,27,Human
+76,0,1,11,Human
 77,1,0,23,Human
-78,0,0,27,Human
-79,1,0,12,Human
+78,1,1,27,Human
+79,1,1,12,Human
 80,0,0,36,Human
-81,1,0,8,Human
-82,1,0,31,Human
-83,0,1,34,Human
+81,1,1,8,Human
+82,0,1,31,Human
+83,1,1,34,Human
 84,0,0,43,Human
-85,0,1,15,Human
-86,1,0,38,Human
+85,0,0,15,Human
+86,1,1,38,Human
 87,1,0,21,Human
-88,1,0,21,Human
-89,1,0,14,Human
-90,0,0,45,Human
-91,1,1,32,Human
-92,0,1,32,Human
-93,0,0,25,Human
+88,1,1,21,Human
+89,0,1,14,Human
+90,1,1,45,Human
+91,0,0,32,Human
+92,1,1,32,Human
+93,1,1,25,Human
 94,1,0,17,Human
-95,1,1,16,Human
-96,1,1,25,Human
-97,0,1,22,Human
+95,1,0,16,Human
+96,1,0,25,Human
+97,1,1,22,Human
 98,0,1,22,Human
-99,0,0,24,Human
+99,1,1,24,Human
